% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{carolus_config}
\alias{carolus_config}
\title{Get the config file containing paths to data}
\usage{
carolus_config(config_file = "~/.carolus/config.yaml")
}
\arguments{
\item{config_file}{The path of the config file}
}
\value{
A nested list of the parsed YAML
}
\description{
If no file is found, a default on is created based on the
template '.carolus_default_dirs'
}
