% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{carolus_dir}
\alias{carolus_dir}
\title{Get a directory specified in the config file}
\usage{
carolus_dir(..., config_file = "~/.carolus/config.yaml")
}
\arguments{
\item{...}{Keys in the config files to the path}
}
\value{
A path to the directory of interest
}
\description{
If no file is found, a default on is created based on the
template '.carolus_default_dirs'
}
