% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_carolus}
\alias{configure_carolus}
\title{Check that the carolus config exists. Create it otherwise}
\usage{
configure_carolus(config_path = "~/.carolus/config.yaml")
}
\arguments{
\item{config_path}{The path of the config file}
}
\description{
If no file is found, a default on is created based on the
template '.carolus_default_dirs'
}
