% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{configure_python}
\alias{configure_python}
\title{Configure the python virtual environment to use}
\usage{
configure_python()
}
\arguments{
\item{envname}{The name of the venv, "carolus" by default}
}
\description{
Uses a venv called "carolus" by default. Installs pandas if not present
in the venv
}
