% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{data}
\name{.carolus_default_dirs}
\alias{.carolus_default_dirs}
\title{The default config file}
\format{
An object of class \code{list} of length 4.
}
\usage{
.carolus_default_dirs
}
\description{
Creates a default config file in the '~/.carolus' directory.
This file contain the paths of the datasets
}
\keyword{datasets}
