% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBA-processed_file_getters.R
\name{get_asv_taxonomy}
\alias{get_asv_taxonomy}
\title{Get the 'asv_taxonomy' files}
\usage{
get_asv_taxonomy(country = "SE", program = NULL)
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}

\item{program}{The taxonomy program to select. Supports 'epang', 'sintax' and 'vsearch'. Use NULL for default taxonomy}
}
\value{
A data frame of the file
}
\description{
Get the 'asv_taxonomy' files
}
