% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IBA_data.R
\name{get_biological_spikes_taxonomy}
\alias{get_biological_spikes_taxonomy}
\title{Get the 'biological_spikes_taxonomy' files}
\usage{
get_biological_spikes_taxonomy(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the 'biological_spikes_taxonomy' files
}
