% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBA-processed_file_getters.R
\name{get_cluster_consensus_taxonomy}
\alias{get_cluster_consensus_taxonomy}
\title{Get the 'cluster_consensus_taxonomy' files}
\usage{
get_cluster_consensus_taxonomy(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the 'cluster_consensus_taxonomy' files
}
