% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBA-processed_file_getters.R
\name{get_cluster_counts}
\alias{get_cluster_counts}
\title{Get the 'cluster_counts' files}
\usage{
get_cluster_counts(country = "SE")

get_cluster_counts(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file

A data frame of the file
}
\description{
Get the 'cluster_counts' files

Get the 'cluster_counts' files
}
