% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBA-raw_file_getters.R
\name{get_CO1_asv_counts}
\alias{get_CO1_asv_counts}
\title{Get the 'CO1_asv_counts' files}
\usage{
get_CO1_asv_counts(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the 'CO1_asv_counts' files
}
