% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IBA_data.R
\name{get_CO1_sequencing_metadata}
\alias{get_CO1_sequencing_metadata}
\title{Get the 'CO1_sequencing_metadata' files}
\usage{
get_CO1_sequencing_metadata(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the 'CO1_sequencing_metadata' files
}
