% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBA-data_helpers.R
\name{get_IBA_data_file}
\alias{get_IBA_data_file}
\title{Get arrow table from a tabular file}
\usage{
get_IBA_data_file(
  fn_prefix,
  data_source,
  arguments = NULL,
  format = "tsv",
  long = FALSE,
  id_header_name = NULL,
  col_header_name = NULL,
  val_header_name = NULL,
  ...
)
}
\arguments{
\item{fn_prefix}{The prefix of the file}

\item{arguments}{Some arguments to paste at the end of the file prefix}

\item{format}{The file ending. Defaults to 'tsv'}
}
\value{
A data frame of the file
}
\description{
Helper function to fetch data for a list of given arguments in the IBA dataset
Files should look like '{fn_prefix}_{arguments}.{format}'
}
