% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBA-processed_file_getters.R
\name{get_processed_dataset}
\alias{get_processed_dataset}
\title{Get the full processed dataset}
\usage{
get_processed_dataset(country = "SE", program = NULL)
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the full processed dataset
}
