% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IBA_data.R
\name{get_raw_file_country}
\alias{get_raw_file_country}
\title{Get arrow table from a tabular file with country data}
\usage{
get_raw_file_country(fn_prefix, country = "SE", format = "tsv")
}
\arguments{
\item{fn_prefix}{The prefix of the file}

\item{country}{The country code. Either 'SE' or 'MG'}

\item{format}{The file ending. Defaults to 'tsv'}
}
\value{
A data frame of the file
}
\description{
Helper function to fetch data for a given country code in the IBA dataset
Files should look like '{fn_prefix}_{country code}.{format}'
}
