% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBA-processed_file_getters.R
\name{get_removed_control_tax}
\alias{get_removed_control_tax}
\title{Get the 'removed_control_tax' files}
\usage{
get_removed_control_tax(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the 'removed_control_tax' files
}
