% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IBA_data.R
\name{get_sample_metadata_litter}
\alias{get_sample_metadata_litter}
\title{Get the metadata litter files}
\usage{
get_sample_metadata_litter(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the metadata litter files
}
