% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IBA_data.R
\name{get_samples_metadata_malaise}
\alias{get_samples_metadata_malaise}
\title{Get the 'samples_metadata_malaise' files}
\usage{
get_samples_metadata_malaise(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the 'samples_metadata_malaise' files
}
