% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBA-raw_file_getters.R
\name{get_soil_chemistry}
\alias{get_soil_chemistry}
\title{Get the 'soil_chemistry' files}
\usage{
get_soil_chemistry(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the 'soil_chemistry' files
}
