% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IBA_data.R
\name{get_stand_characteristics_MG}
\alias{get_stand_characteristics_MG}
\title{Get the 'stand_characteristic' file. Madagascar specific}
\usage{
get_stand_characteristics_MG()
}
\value{
A data frame of the file
}
\description{
Get the 'stand_characteristic' file. Madagascar specific
}
