% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IBA_data.R
\name{get_synthetic_spikes_info}
\alias{get_synthetic_spikes_info}
\title{Get the 'synthetic_spikes_info' file.}
\usage{
get_synthetic_spikes_info()
}
\value{
A data frame of the file
}
\description{
Get the 'synthetic_spikes_info' file.
}
