% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon-search.R
\name{taxon_search}
\alias{taxon_search}
\title{Search the dataset for a set of taxa}
\usage{
taxon_search(taxa, rank, treatment_type = "all", dataset = "IBA_SE")
}
\arguments{
\item{taxa}{A character vector containing the taxa of interest}

\item{rank}{The rank of the taxa supplied}

\item{treatment_type}{The treatment type of the sample: 'lysate', 'homogenate' or 'all'}

\item{dataset}{The dataset name: 'IBA_SE' or 'IBA_MG'}
}
\value{
A data frame containing the sample ids of the taxa with metadata
}
\description{
Search the dataset for a set of taxa
}
\examples{
# Search for a single taxon 
taxon_search("Trypoxylon attenuatum", "Species")

# Search for a multiple taxa
taxon_search(c("Astata", "Mellinus", "Trypoxylon"), "Genus")

# Search for lysate samples
taxon_search("Astata boops", "Species", treatment_type = "lysate")

}
