% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_IBA_data.R
\name{get_biomass_count}
\alias{get_biomass_count}
\title{Get the 'biological_spikes_taxonomy' files}
\usage{
get_biomass_count(dataset = "IBA")
}
\arguments{
\item{dataset}{The dataset name. Either 'IBA' or 'SIIP'}
}
\value{
A data frame of the file
}
\description{
Get the 'biological_spikes_taxonomy' files
}
