% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBA-raw_file_getters.R
\name{get_sites_metadata}
\alias{get_sites_metadata}
\title{Get the 'sites_metadata' files}
\usage{
get_sites_metadata(country = "SE")
}
\arguments{
\item{country}{The country code. Either 'SE' or 'MG', defaults to 'SE'}
}
\value{
A data frame of the file
}
\description{
Get the 'sites_metadata' files
}
